requirements.txt
Flask==1.1.1
requests==2.24.0
ptvsd==4.3.2 # Required for debugging.
google-cloud-bigquery
looker_sdk==0.1.3b15